tic_tac_toe: add unit tests
